Koios 1.0.6 Alignment (#56)
* Fix TxHash Object

* Add Javadoc

* logging-interceptor version bump to 4.9.3
bcprov-jdk15to18 version bump to 1.71
maven-javadoc-plugin version bump to 3.40
maven-gpg-plugin version bump to 3.0.1

* Bump Koios Java Client Version to 1.11

* Vasil HF Changes Preparations

* Fix Requests with One Option Filter

* Added BlockTxHash Javadoc

* Bump logging-interceptor version 4.9.3 -> 4.10.0 due to a vulnerability

* Epoch Params CostModels Object Type Change from String to Map

* BaseService Initialization Logic Improvement + Inner Retry Mechanism upon Timeouts + Timeouts and Retries Env Variables Support

* Bump junit-jupiter version 5.8.2 -> 5.9.0

* Upgrade CodeQL Workflow to v2

* Jacoco to Generate Coverage Endpoint

* Cosmetics

* Javadoc

* Epoch Fields from Long to Int (Breaking Change)

* Datetime Fields to Unix Timestamp (String -> Long)

* Lovelace fields Data type changed to String instead of Long

* Javadoc

* Koios 1.0.6 Alignment

+++ src/main/java/rest/koios/client/backend/api/epoch/EpochService.java
@@ -36,7 +36,7 @@ public interface EpochService {
      * @return Result of {@link EpochInfo} detailed summary
      * @throws ApiException if an error occurs while attempting to invoke the API
      */
-    Result<EpochInfo> getEpochInformationByEpoch(Long epochNo) throws ApiException;
+    Result<EpochInfo> getEpochInformationByEpoch(Integer epochNo) throws ApiException;
 
     /**
      * Epoch Information
@@ -74,7 +74,7 @@ public interface EpochService {
      * @return Result of {@link EpochParams} protocol parameters
      * @throws ApiException if an error occurs while attempting to invoke the API
      */
-    Result<EpochParams> getEpochParametersByEpoch(Long epochNo) throws ApiException;
+    Result<EpochParams> getEpochParametersByEpoch(Integer epochNo) throws ApiException;
 
     /**
      * Epoch's Protocol Parameters
